moves file-reading functions to a new, separate file
