+++ src/writer.rs
@@ -1,4 +1,3 @@
-// use fgift::gets;
 use std::fs::File;
 use std::fs::OpenOptions;
 use std::io;
